pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "projinit"
version = "0.1.0"
description = "CLI pour générer la structure d'un projet avec configuration Terraform GitHub"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Xavier Gueret", email = "xgueret@users.noreply.github.com" }
]
keywords = ["cli", "terraform", "github", "project-generator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "questionary>=2.0.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
projinit = "projinit.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/projinit"]
